+++ resources/views/Components/tableAddNew.blade.php
@@ -1,7 +1,0 @@
-<link rel="stylesheet" href="{{ asset('css/TableAddNew.css') }}">
-
-<div class="table-add-new">
-    <div class="table-add-newDropDown">
-        <button class="table-add-new-DropBtn" id="myBtn">Изменить</button>
-    </div>
-</div>