+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Development
 
+* 2024-06-28 -- 1.1.6 - Images upload
+  * Image upload added
+  * Removed unnecessary files
+  * Translations fixed 
+
 * 2024-06-28 -- 1.1.5 - Laravel Voyager
   * Added Laravel Voyager
   * Added Laravel Voyager Bread Seeder